dev/services/python-service/requirements.txt
uvicorn==0.18.3
fastapi==0.85.0
redis==4.3.4
loguru==0.6.0
PyYAML==6.0
